+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed `Client.clear_commands` to `Client.clear_application_commands`.
 - Renamed `declare_slash_command` and `declare_slash_commands` to `declare_application_command` and
   `declare_application_commands` respectively.
+- Renamed `Client.detect_commands` to `Client.load_from_scope`.
+- Restructured LoadableProtocol for re-use in `Client.load_from_scope` and rename to `ComponentLoader`.
 
 ### Fixed
 - Don't include the "tracked command ID" in slash command group builders as this leads to mis-matching ID
@@ -30,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BaseSlashCommand.tracked_command_id is no-longer used in command builders and cannot passed to
   `as_slash_command`, `slash_command_group`, `SlashCommand.__init__` and `SlashCommandGroup.__init__`
   as `command_id` anymore.
+- `load_from_attributes` behaviour from the standard Component implementation.
 
 ## [2.1.4a1] - 2021-11-15
 ### Added